requirements.txt
absl-py==2.3.1
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autograd==1.8.0
autograd-gamma==0.5.0
biopython==1.85
blinker==1.9.0
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools==6.2.0
certifi @ file:///C:/b/abs_67mh6i36ci/croot/certifi_1752653721535/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
flatbuffers==25.2.10
fonttools==4.59.1
formulaic==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httptools==0.6.4
huggingface-hub==0.34.4
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
iniconfig==2.1.0
interface-meta==1.3.0
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
lifelines==0.30.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.5
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.3
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.6.3
namex==0.1.0
narwhals==2.1.2
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numexpr @ file:///C:/b/abs_4e0h6i5ip2/croot/numexpr_1752521729801/work
numpy==2.2.6
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas @ file:///C:/b/abs_d5idbck212/croot/pandas_1752872294545/work/dist/pandas-2.3.1-cp310-cp310-win_amd64.whl#sha256=aa0d0a8165265b53af6bf391a0e2723d1073ac22447a5bc23ad6e9245c74bfda
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyg-lib==0.4.0+pt27cu118
Pygments==2.19.2
pyparsing==3.2.3
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
pytest==8.4.1
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-dotenv==1.1.1
python-multipart==0.0.20
pytz @ file:///C:/b/abs_f8wdzeix0n/croot/pytz_1752135878094/work
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
rdkit-pypi==2022.9.5
referencing==0.36.2
regex==2025.7.34
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
rich==14.1.0
rpds-py==0.27.1
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.3
streamlit==1.50.0
streamlit-drawable-canvas==0.9.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
torch==2.7.1+cu118
torch-geometric==2.6.1
torch_cluster==1.6.3+pt27cu118
torch_scatter==2.1.2+pt27cu118
torch_sparse==0.6.18+pt27cu118
torch_spline_conv==1.2.2+pt27cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
transformers==4.55.0
typing-inspection==0.4.1
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wrapt==1.17.3
yarl==1.20.1
